Move impl.variable.* to impl.stats.* because variable-code is statistics-code
