+++ probdiffeq/impl/_stats.py
@@ -28,6 +28,14 @@ class StatsBackend(abc.ABC):
     def transform_unit_sample(self, unit_sample, /, rv):
         raise NotImplementedError
 
+    @abc.abstractmethod
+    def to_multivariate_normal(self, u, rv):
+        raise NotImplementedError
+
+    @abc.abstractmethod
+    def rescale_cholesky(self, rv, factor, /):
+        raise NotImplementedError
+
 
 class ScalarStats(StatsBackend):
     def mahalanobis_norm_relative(self, u, /, rv):
@@ -59,6 +67,14 @@ class ScalarStats(StatsBackend):
 
     def transform_unit_sample(self, unit_sample, /, rv):
         return rv.mean + rv.cholesky @ unit_sample
+
+    def rescale_cholesky(self, rv, factor):
+        if np.ndim(factor) > 0:
+            return functools.vmap(self.rescale_cholesky)(rv, factor)
+        return _normal.Normal(rv.mean, factor * rv.cholesky)
+
+    def to_multivariate_normal(self, rv):
+        return rv.mean, rv.cholesky @ rv.cholesky.T
 
 
 class DenseStats(StatsBackend):
@@ -99,6 +115,13 @@ class DenseStats(StatsBackend):
 
     def transform_unit_sample(self, unit_sample, /, rv):
         return rv.mean + rv.cholesky @ unit_sample
+
+    def rescale_cholesky(self, rv, factor, /):
+        cholesky = factor[..., None, None] * rv.cholesky
+        return _normal.Normal(rv.mean, cholesky)
+
+    def to_multivariate_normal(self, rv):
+        return rv.mean, rv.cholesky @ rv.cholesky.T
 
 
 class IsotropicStats(StatsBackend):
@@ -145,6 +168,17 @@ class IsotropicStats(StatsBackend):
     def transform_unit_sample(self, unit_sample, /, rv):
         return rv.mean + rv.cholesky @ unit_sample
 
+    def rescale_cholesky(self, rv, factor, /):
+        cholesky = factor[..., None, None] * rv.cholesky
+        return _normal.Normal(rv.mean, cholesky)
+
+    def to_multivariate_normal(self, rv):
+        eye_d = np.eye(*self.ode_shape)
+        cov = rv.cholesky @ rv.cholesky.T
+        cov = np.kron(eye_d, cov)
+        mean = rv.mean.reshape((-1,), order="F")
+        return (mean, cov)
+
 
 class BlockDiagStats(StatsBackend):
     def __init__(self, ode_shape):
@@ -185,3 +219,17 @@ class BlockDiagStats(StatsBackend):
 
     def transform_unit_sample(self, unit_sample, /, rv):
         return rv.mean + (rv.cholesky @ unit_sample[..., None])[..., 0]
+
+    def rescale_cholesky(self, rv, factor, /):
+        cholesky = factor[..., None, None] * rv.cholesky
+        return _normal.Normal(rv.mean, cholesky)
+
+    def to_multivariate_normal(self, rv):
+        mean = np.reshape(rv.mean.T, (-1,), order="F")
+        cov = np.block_diag(self._cov_dense(rv.cholesky))
+        return (mean, cov)
+
+    def _cov_dense(self, cholesky):
+        if cholesky.ndim > 2:
+            return functools.vmap(self._cov_dense)(cholesky)
+        return cholesky @ cholesky.T
